[None][chore] Move top-p decay handling into top_p_decay.py
The feature's per-slot store and its whole lifecycle (admission,
pre-sample metadata, post-sample recurrence, retirement) move out of
TorchSampler into a TopPDecayHandler in a dedicated submodule, mirroring
how token_ban.py is organized.

The handler owns both the TopPDecayStore buffers and the host-side
membership set that gates them, so the two can no longer drift: the set
was previously a TorchSampler attribute while the store hung off
TorchSampler.Store. TorchSampler keeps a single _top_p_decay attribute
and drives it through the four lifecycle hooks.

update_after_sample now takes the already-indexed per-slot step tokens
instead of the full new_tokens tensor, keeping the decay-is-single-token
assumption at the call site. No behavior change.
